@@ -49,72 +51,71 @@
     "babel-register": "^6.26.0",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.17.2",
-    "cloudinary": "^1.11.0",
-    "cloudinary-core": "^2.5.0",
-    "cloudinary-react": "^1.0.6",
+    "cloudinary": "^1.28.1",
+    "cloudinary-core": "^2.12.3",
+    "cloudinary-react": "^1.7.1",
     "connect-history-api-fallback": "^1.3.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^1.0.0",
+    "copy-webpack-plugin": "^10.2.4",
+    "css-loader": "^6.6.0",
     "dotenv": "^6.2.0",
     "dotenvenc": "^1.0.4",
-    "express": "^4.15.4",
-    "express-winston": "^3.0.1",
+    "express": "^4.17.3",
+    "express-winston": "^4.2.0",
     "flat-cache": "^1.3.4",
-    "html-webpack-plugin": "^3.2.0",
-    "jquery": "^3.3.1",
+    "html-webpack-plugin": "^5.5.0",
+    "jquery": "^3.6.0",
     "jsonwebtoken": "^8.3.0",
-    "lodash": "^4.17.5",
-    "material-ui": "^1.0.0-beta.41",
-    "mini-css-extract-plugin": "^0.4.3",
+    "lodash": "^4.17.21",
+    "mini-css-extract-plugin": "^2.5.3",
     "moment": "^2.22.1",
-    "mongoose": "^5.3.0",
+    "mongoose": "^6.2.2",
     "morgan": "^1.9.1",
-    "node-sass": "^4.10.0",
+    "node-sass": "^7.0.1",
     "nodemailer": "^4.6.7",
     "nodemon": "^1.11.0",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
     "progress-bar-webpack-plugin": "^1.11.0",
-    "prop-types": "^15.6.2",
+    "prop-types": "^15.8.1",
     "query-string": "^6.2.0",
-    "react": "^16.6.3",
+    "react": "^17.0.2",
     "react-anchor-link-smooth-scroll": "^1.0.11",
     "react-animate-on-scroll": "^2.1.5",
-    "react-bootstrap": "^0.32.1",
+    "react-bootstrap": "^2.1.2",
     "react-count-to": "^0.12.0",
     "react-dev-utils": "^5.0.2",
     "react-device-detect": "^1.6.1",
-    "react-dom": "^16.2",
+    "react-dom": "^17.0.2",
     "react-ga": "^2.5.3",
     "react-google-maps": "^9.4.5",
     "react-hot-loader": "^4.3.11",
-    "react-icons": "^3.2.2",
-    "react-images": "^0.5.17",
+    "react-icons": "^4.3.1",
+    "react-images": "^1.2.0-beta.7",
     "react-magnific-popup": "^1.0.1",
-    "react-quill": "^1.3.1",
+    "react-quill": "^1.3.5",
     "react-router": "^4.1.2",
     "react-router-dom": "^4.1.2",
-    "react-toastify": "^4.1.0",
-    "react-tooltip": "^3.9.2",
-    "react-typed": "^1.0.8",
-    "react-youtube": "^7.6.0",
+    "react-toastify": "^8.2.0",
+    "react-tooltip": "^4.2.21",
+    "react-typed": "^1.2.0",
+    "react-youtube": "^7.14.0",
     "recompose": "^0.30.0",
-    "sanitize-html": "^1.18.3",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.0",
-    "styled-components": "^4.1.3",
-    "terser-webpack-plugin": "^1.2.1",
+    "sanitize-html": "^2.7.0",
+    "sass-loader": "^12.6.0",
+    "style-loader": "^3.3.1",
+    "styled-components": "^5.3.3",
+    "terser-webpack-plugin": "^5.3.1",
     "transform-runtime": "0.0.0",
-    "uglify-js": "^3.4.9",
+    "uglify-js": "^3.15.1",
     "uglify-js2": "^2.1.11",
-    "uglifyjs-webpack-plugin": "^2.0.1",
-    "webpack": "^4.20.2",
-    "webpack-dev-middleware": "^3.4.0",
-    "webpack-hot-middleware": "^2.18.2",
-    "webpack-merge": "^4.1.0",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "webpack": "^5.69.1",
+    "webpack-dev-middleware": "^5.3.1",
+    "webpack-hot-middleware": "^2.25.1",
+    "webpack-merge": "^5.8.0",
     "whatwg-fetch": "^3.0.0",
     "winston": "^3.0.0",
-    "yarn": "^1.13.0"
+    "yarn": "^1.22.17"
   },
   "devDependencies": {
     "babel-eslint": "^10.0.1",
