@@ -1,8 +1,8 @@
 import React from 'react';
-import QueueMusic from '@material-ui/icons/QueueMusic';
-import Work from '@material-ui/icons/Work';
-import TagFaces from '@material-ui/icons/TagFaces';
-import School from '@material-ui/icons/School';
+import QueueMusic from '@mui/icons-material/QueueMusic';
+import Work from '@mui/icons-material/Work';
+import TagFaces from '@mui/icons-material/TagFaces';
+import School from '@mui/icons-material/School';
 import get from 'lodash/get';
 import data from './servicesData.json';
 import ServiceBox from './ServiceBox';
