@@ -1,4 +1,4 @@
-const merge = require('webpack-merge');
+const { merge } = require('webpack-merge');
 
 const commonConfig = require('./webpack.common');
 
@@ -21,5 +21,6 @@ module.exports = merge(commonConfig, {
     contentBase: './client/public',
     historyApiFallback: true,
     stats: 'minimal', // none (or false), errors-only, minimal, normal (or true) and verbose
+    writeToDisk: true,
   },
 });
